refactor(vnui): refactor theme
