+++ applications/tari_validator_node_web_ui/src/assets/Logo.tsx
