+++ applications/tari_validator_node_web_ui/src/index.tsx
@@ -22,7 +22,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './theme/theme.css';
 import reportWebVitals from './reportWebVitals';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import ValidatorNode from './routes/VN/ValidatorNode';
