+++ applications/tari_validator_node_web_ui/src/theme/fonts/AvenirLTStd-Heavy.otf
